Move SlimEnumDropdown into it's own file
