+++ osu.Game/Overlays/Direct/FilterControl.cs
@@ -10,7 +10,6 @@ using osu.Framework.Extensions.Color4Extensions;
 using osu.Framework.Graphics;
 using osu.Framework.Graphics.Containers;
 using osu.Framework.Graphics.Sprites;
-using osu.Framework.Graphics.UserInterface;
 using osu.Game.Database;
 using osu.Game.Graphics;
 using osu.Game.Graphics.Sprites;
@@ -283,38 +282,6 @@ namespace osu.Game.Overlays.Direct
             protected override void Dispose(bool isDisposing)
             {
                 bindable.ValueChanged -= Bindable_ValueChanged;
-            }
-        }
-
-        private class SlimEnumDropdown<T> : OsuEnumDropdown<T>
-        {
-            public const float HEIGHT = 25;
-
-            protected override DropdownHeader CreateHeader() => new SlimDropdownHeader { AccentColour = AccentColour };
-            protected override Menu CreateMenu() => new SlimMenu();
-
-            private class SlimDropdownHeader : OsuDropdownHeader
-            {
-                public SlimDropdownHeader()
-                {
-                    Height = HEIGHT;
-                    Icon.TextSize = 16;
-                    Foreground.Padding = new MarginPadding { Top = 4, Bottom = 4, Left = 8, Right = 4 };
-                }
-
-                protected override void LoadComplete()
-                {
-                    base.LoadComplete();
-                    BackgroundColour = Color4.Black.Opacity(0.25f);
-                }
-            }
-
-            private class SlimMenu : OsuMenu
-            {
-                public SlimMenu()
-                {
-                    Background.Colour = Color4.Black.Opacity(0.25f);
-                }
             }
         }
     }
